+++ src/main/java/com/chrisgruber/nanourl/repositories/UrlAliasRepository.java
@@ -4,6 +4,6 @@ import com.chrisgruber.nanourl.models.UrlAlias;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface UrlAliasRepository extends MongoRepository<UrlAlias, String> {
-    UrlAlias findById(int id);
+    UrlAlias findById(long id);
     UrlAlias findByAliasPath(String aliasPath);
 }
